Debugged plotting implementation w.r.t. imports, files; added tests
