+++ src/simmer/drivers.py
@@ -9,8 +9,7 @@ import darks
 import flats
 import image
 import pandas as pd
-import plotting
-import schemas.read_yml
+import plotting as pl
 import sky
 from tqdm import tqdm
 
@@ -33,7 +32,8 @@ def all_driver(inst, config_file, raw_dir, reddir, plotting_yml=None):
     config = pd.read_csv(config_file)
     config.Object = config.Object.astype(str)
 
-    plotting.plot_config = get_plotting_args(plotting_yml)
+    if plotting_yml:
+        pl.initialize_plotting(plotting_yml)
 
     darks.dark_driver(raw_dir, reddir, config, inst)
     flats.flat_driver(raw_dir, reddir, config, inst)
