+++ src/simmer/schemas/plotting.yml
@@ -1,6 +1,4 @@
 # For plotting purposes. Descriptions provided as meta.
-# TODO: make sure that the "list validation" works — or add to custom validator
-# TODO: add yaml linter to checks
 final_im:
   type: "dict"
   schema:
@@ -15,6 +13,7 @@ final_im:
         to the .png file.
     zoom_scale:
       type: "integer"
+      'forbidden': [true]
       min: 0
       default: 0
       meta: zoom in on the center of an image with box width of `zoom`
@@ -35,30 +34,65 @@ final_im:
       default: "linear"
       meta: Determines the scaling of the color. Can be linear, square root,
         log (base 10), sine, or exponential.
-    matplotlib_kwargs:
-      type: "string"
 
 rots:
   type: "dict"
   schema:
     plot:
       type: "boolean"
-      default: "true"
+      default: true
       meta: Whether or not to produce a plot of the final image.
     output_fits:
       type: "boolean"
-      default: "true"
+      default: true
       meta: Whether or not to output a .fits file of the final image in addition
         to the .png file.
     zoom_scale:
       type: "integer"
-      min: "1"
-      default: "0"
+      'forbidden': [true]
+      min: 0
+      default: 0
       meta: zoom in on the center of an image with box width of `zoom`
         pixels. A zoom of 0 leads to no zoom.
     colorbars:
       type: "boolean"
-      default: "true"
+      default: true
+      meta: Setting to determine whether to turn on colorbars (labeled) for
+        all plots produced by SImMER.
+    colormap:
+      type: "string"
+      default: "plasma"
+      meta: Sets the colormap of all produced plots produced by SImMer.
+        Checks against matplotlib colormaps later on and will default
+        to "plasma".
+    scaling:
+      type: "string"
+      default: "linear"
+      meta: Determines the scaling of the color. Can be linear, square root,
+        log (base 10), sine, or exponential.
+
+intermediate:
+  type: "dict"
+  schema:
+    plot:
+      type: "boolean"
+      default: true
+      meta: Whether or not to produce a plot of the final image.
+    output_fits:
+      type: "boolean"
+      default: true
+      meta: Whether or not to output a .fits file of the final image in addition
+        to the .png file.
+    zoom_scale:
+      type: "integer"
+      'forbidden': [true]
+      min: 0
+      default: 0
+      meta: zoom in on the center of an image with box width of `zoom`
+        pixels. A zoom of 0 leads to no zoom.
+    colorbars:
+      type: "boolean"
+      default: true
       meta: Setting to determine whether to turn on colorbars (labeled) for
         all plots produced by SImMER.
     colormap:
